Fix profiling review policy issues
