+++ Sources/App/MenuBarProfilingLauncher.swift
@@ -51,25 +51,3 @@ enum MenuBarProfilingLauncher {
         }
     }
 }
-
-@MainActor
-enum MenuBarProfilingMenuItem {
-    static func make() -> NSMenuItem {
-        let item = NSMenuItem(
-            title: String(localized: "statusMenu.startProfiling", defaultValue: "Start Profiling"),
-            action: #selector(MenuBarProfilingMenuItemTarget.startProfiling(_:)),
-            keyEquivalent: ""
-        )
-        item.target = MenuBarProfilingMenuItemTarget.shared
-        return item
-    }
-}
-
-@MainActor
-private final class MenuBarProfilingMenuItemTarget: NSObject {
-    static let shared = MenuBarProfilingMenuItemTarget()
-
-    @objc func startProfiling(_ sender: NSMenuItem) {
-        MenuBarProfilingLauncher.start()
-    }
-}
